+++ cookbooks/devel/recipes/vagrant.rb
@@ -19,6 +19,10 @@
 #
 
 # refs:
+#   https://docs.vagrantup.com/v2/synced-folders/nfs.html
+#   https://github.com/mitchellh/vagrant/issues/1941#issuecomment-68608908
+#   http://askubuntu.com/questions/103910/nfs-is-blocked-by-ufw-even-though-ports-are-opened
+#   https://github.com/genesis/wordpress#ssh---ssh-authentication-failed
 #   https://github.com/smdahlen/vagrant-hostmanager
 
 devel = node[:apps][:devel]
@@ -32,12 +36,45 @@ if app_available? vagrant
     profile vagrant
   end
 
+  # NFS support
+  %w(
+    libgssglue1
+    libnfsidmap2
+    libtirpc1
+    nfs-common
+    nfs-kernel-server
+    rpcbind
+  ).each do |pkg|
+    package pkg
+  end
+
   cookbook_file '/etc/sudoers.d/10_vagrant_nfs' do
     source '/vagrant/vagrant_nfs'
     mode 00440
     backup false
   end
 
+  cookbook_file 'nfs-kernel-server' do
+    path '/etc/default/nfs-kernel-server'
+    source '/vagrant/nfs-kernel-server'
+    mode 0644
+    backup false
+  end
+
+  cookbook_file 'ufw_nfs' do
+    path '/etc/ufw/applications.d/nfs'
+    source '/vagrant/ufw_nfs'
+    mode 0644
+    backup false
+  end
+
+  execute 'ufw_allow_nfs' do
+    command 'ufw allow NFS'
+    action :nothing
+    subscribes :run, resources('cookbook_file[ufw_nfs]'), :immediately
+  end
+
+  # Vagrant insecure public key
   cache_path = Chef::Config[:file_cache_path]
 
   cookbook_file "#{cache_path}/vagrant.pub" do
@@ -60,9 +97,7 @@ if app_available? vagrant
         # Loading user environment ...
         . ${HOME}/.bashrc
 
-        [[ -f ${HOME}/.vagrant.d/insecure_private_key ]] && ssh-add ${HOME}/.vagrant.d/insecure_private_key
         cat #{cache_path}/vagrant.pub >> ${HOME}/.ssh/authorized_keys
-
         vagrant plugin install vagrant-hostmanager
         EOH
       action :run